LayoutTests/fast/encoding/char-decoding-mac.html

<html>
<head>
<script src="../../resources/js-test.js"></script>
<script src="resources/char-decoding-utils.js"></script>
</head>
<body>
<script>

description("This tests decoding characters in various character sets.");

// Legacy Mac encodings.
testDecode('mac', '%C3', 'U+221A');
testDecode('macintosh', '%C3', 'U+221A');
testDecode('MacRoman', '%C3', 'U+221A');
testDecode('x-mac-greek', '%B0', 'U+FFFD');
testDecode('x-mac-ukrainian', '%80', 'U+0410');
testDecode('x-mac-cyrillic', '%80', 'U+0410');
testDecode('mac-cyrillic', '%80', 'U+0410');
testDecode('x-mac-centraleurroman', '%81', 'U+FFFD');
testDecode('x-mac-ce', '%81', 'U+FFFD');
testDecode('x-mac-turkish', '%81', 'U+FFFD');

</script>
</body>
</html>
